+++ src/config/ava.js
@@ -10,7 +10,7 @@ module.exports = {
   rules: {
     'ava/assertion-arguments': 'error',
     'ava/assertion-message': ['off', 'always'],
-    'ava/hooks-order': 'off',
+    'ava/hooks-order': 'error',
     'ava/max-asserts': ['off', 5],
     'ava/no-async-fn-without-await': 'error',
     'ava/no-cb-test': 'off',